moved more logic out of container
