+++ src/pages/ProblemEditor/ProblemEditorContainer/ExampleValidator.tsx
@@ -10,14 +10,27 @@ import {
   validateTestEditor,
 } from "./problemEditorContainerSlice";
 
-interface Props {
-  validationStatus: () => boolean;
-}
-
-export const ExampleValidator = ({ validationStatus }: Props) => {
+export const ExampleValidator = () => {
   const dispatch = useDispatch();
   const activeStep = useAppSelector(
     (state) => state.problemEditorContainer.activeStep
+  );
+
+  /* STEP SELECTORS */ //TODO put this logic in the reducer
+  const metadataIsValid = useAppSelector(
+    (state) => state.problemEditorContainer.metadataIsValid
+  );
+  const problemIsValid = useAppSelector(
+    (state) => state.problemEditorContainer.problemIsValid
+  );
+  const codeIsValid = useAppSelector(
+    (state) => state.problemEditorContainer.codeIsValid
+  );
+  const serverConfigIsValid = useAppSelector(
+    (state) => state.problemEditorContainer.serverConfigIsValid
+  );
+  const testEditorIsValid = useAppSelector(
+    (state) => state.problemEditorContainer.testEditorIsValid
   );
 
   /* Dispatches the correct action based on the current active step */
@@ -41,6 +54,24 @@ export const ExampleValidator = ({ validationStatus }: Props) => {
     }
   };
 
+  /* Selects the correct member of state based on the current active step */
+  const getValidationStatus = (): boolean => {
+    switch (activeStep) {
+      case 0:
+        return metadataIsValid;
+      case 1:
+        return problemIsValid;
+      case 2:
+        return codeIsValid;
+      case 3:
+        return serverConfigIsValid;
+      case 4:
+        return testEditorIsValid;
+      default:
+        return false;
+    }
+  };
+
   return (
     <Box mt="auto" ml="auto" mr="auto">
       <FormGroup>
@@ -49,7 +80,7 @@ export const ExampleValidator = ({ validationStatus }: Props) => {
             <Switch
               color="primary"
               onChange={(event) => handleValidation(event.target.checked)}
-              checked={validationStatus()}
+              checked={getValidationStatus()}
             />
           }
           label="Validated"
